Reformat welcome text to be button and modal
Since creating the welcome message, the Modal component was
created. It made more sense to keep the welcome message in the
Modal component, as it would be a modal itself. To keep the logic
consistent with other modal components, emptyListModal local state
was created and tied to an button click listener to dictate modal
display/ not display. The local state was also integral to be able
to switch the button message depending on if the modal was open
or closed.